requirements.txt
python-telegram-bot==21.6
garminconnect==0.2.8
httpx==0.27.2
sqlalchemy==2.0.36
aiosqlite==0.20.0
anthropic==0.40.0
python-dotenv==1.0.1
aiohttp==3.11.10
tenacity==9.0.0
cryptography==43.0.3
